additional_requirements.txt
absl-py==0.8.0
appdirs==1.4.3
appnope==0.1.0
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
blis==0.2.4
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
cymem==2.0.2
decorator==4.4.0
defusedxml==0.6.0
dlib==19.18.0
docker==4.0.2
entrypoints==0.3
falcon==2.0.0
gast==0.3.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.10.0
idna==2.8
imageio==2.5.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
murmurhash==1.0.2
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.5
notebook==6.0.1
numpy==1.17.2
opencv-python==4.1.1.26
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
plac==0.9.6
preshed==2.0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.2
ptyprocess==0.6.0
Pygments==2.4.2
pymongo==3.9.0
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
pytools==2019.1.1
pytz==2019.2
PyWavelets==1.0.3
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.22.0
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
Send2Trash==1.5.0
six==1.12.0
spacy==2.1.8
srsly==0.1.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
thinc==7.0.8
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.2
urllib3==1.25.5
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
